+++ sdk/individualAccountAttachment/individualAccountAttachment.js
@@ -3,32 +3,32 @@ const check = require('starkcore').check;
 const Resource = require('starkcore').Resource;
 
 
-class AccountRequestAttachment extends Resource {
+class IndividualAccountAttachment extends Resource {
     /**
      *
-     * AccountRequestAttachment object
+     * IndividualAccountAttachment object
      *
-     * @description You can create an Account Request Attachment to attach images of documents
+     * @description You can create an Individual Account Attachment to attach images of documents
      * to a specific Individual Account Request. You must reference the desired Individual Account Request by its id.
-     * 
-     * When you initialize a AccountRequestAttachment, the entity will not be automatically
+     *
+     * When you initialize a IndividualAccountAttachment, the entity will not be automatically
      * created in the Stark Infra API. The 'create' function sends the objects
      * to the Stark Infra API and returns the list of created objects.
      *
      * Parameters (required):
-     * @param type [string]: type of the AccountRequestAttachment. Options: 'drivers-license-front', 'drivers-license-back', 'identity-front', 'identity-back' or 'selfie'
+     * @param type [string]: type of the IndividualAccountAttachment. Options: 'drivers-license-front', 'drivers-license-back', 'identity-front', 'identity-back' or 'selfie'
      * @param content [string]: Base64 data url of the picture. ex: data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAASABIAAD...
      * @param accountRequestId [string]: Unique id of IndividualAccountRequest. ex: '5656565656565656'
      * @param contentType [string]: content MIME type. This parameter is required as input only. ex: 'image/png' or 'image/jpeg'
-     * 
+     *
      * Parameters (optional):
-     * @param tags [list of strings, default []]: list of strings for reference when searching for AccountRequestAttachments. ex: ['employees', 'monthly']
-     * 
+     * @param tags [list of strings, default []]: list of strings for reference when searching for IndividualAccountAttachments. ex: ['employees', 'monthly']
+     *
      * Attributes (return-only):
-     * @param id [string]: unique id returned when AccountRequestAttachment is created. ex: '5656565656565656'
-     * @param status [string]: current status of the AccountRequestAttachment. ex: 'created'
-     * @param created [string]: creation datetime for the AccountRequestAttachment. ex: '2020-03-10 10:30:00.000'
-     * 
+     * @param id [string]: unique id returned when IndividualAccountAttachment is created. ex: '5656565656565656'
+     * @param status [string]: current status of the IndividualAccountAttachment. ex: 'created'
+     * @param created [string]: creation datetime for the IndividualAccountAttachment. ex: '2020-03-10 10:30:00.000'
+     *
      */
     constructor({
         type, content = null, accountRequestId, contentType, created = null, status = null, id = null
@@ -50,24 +50,24 @@ class AccountRequestAttachment extends Resource {
     }
 }
 
-exports.AccountRequestAttachment = AccountRequestAttachment;
-let resource = {'class': exports.AccountRequestAttachment, 'name': 'AccountRequestAttachment'};
+exports.IndividualAccountAttachment = IndividualAccountAttachment;
+let resource = {'class': exports.IndividualAccountAttachment, 'name': 'IndividualAccountAttachment'};
 
 exports.create = async function (attachments, { user } = {}) {
     /**
      *
-     * Create AccountRequestAttachments
+     * Create IndividualAccountAttachments
      *
-     * @description Send a list of AccountRequestAttachment objects for creation in the Stark Infra API
+     * @description Send a list of IndividualAccountAttachment objects for creation in the Stark Infra API
      *
      * Parameters (required):
-     * @param attachments [list of AccountRequestAttachment objects]: list of AccountRequestAttachment objects to be created in the API
+     * @param attachments [list of IndividualAccountAttachment objects]: list of IndividualAccountAttachment objects to be created in the API
      *
      * Parameters (optional):
      * @param user [Organization/Project object]: Organization or Project object. Not necessary if starkinfra.user was set before function call
      *
      * Return:
-     * @returns list of AccountRequestAttachment objects with updated attributes
+     * @returns list of IndividualAccountAttachment objects with updated attributes
      *
      */
     return rest.post(resource, attachments, user);
@@ -76,9 +76,9 @@ exports.create = async function (attachments, { user } = {}) {
 exports.get = async function (id, { user } = {}) {
     /**
      *
-     * Retrieve a specific AccountRequestAttachment
+     * Retrieve a specific IndividualAccountAttachment
      *
-     * @description Receive a single AccountRequestAttachment object previously created in the Stark Infra API by passing its id
+     * @description Receive a single IndividualAccountAttachment object previously created in the Stark Infra API by passing its id
      *
      * Parameters (required):
      * @param id [string]: object unique id. ex: '5656565656565656'
@@ -87,7 +87,7 @@ exports.get = async function (id, { user } = {}) {
      * @param user [Organization/Project object]: Organization or Project object. Not necessary if starkinfra.user was set before function call
      *
      * Return:
-     * @returns AccountRequestAttachment object with updated attributes
+     * @returns IndividualAccountAttachment object with updated attributes
      *
      */
     return rest.getId(resource, id, user);
@@ -96,21 +96,21 @@ exports.get = async function (id, { user } = {}) {
 exports.query = async function ({ limit, after, before, status, tags, ids, user } = {}) {
     /**
      *
-     * Retrieve AccountRequestAttachments
+     * Retrieve IndividualAccountAttachments
      *
-     * @description Receive a generator of AccountRequestAttachment objects previously created in the Stark Infra API and the cursor to the next page. 
+     * @description Receive a generator of IndividualAccountAttachment objects previously created in the Stark Infra API and the cursor to the next page.
      *
      * Parameters (optional):
      * @param limit [integer, default null]: maximum number of objects to be retrieved. Unlimited if null. ex: 35
      * @param after [string, default null]: date filter for objects created after this date. ex: '2020-03-10'
      * @param before [string, default null]: date filter for objects created before this date. ex: '2020-03-10'
      * @param status [string, default null]: filter for status of the retrieved objects. ex: 'created', 'canceled', 'processing', 'failed', 'success'
-     * @param tags [list of strings, default null]: list of strings for reference when searching for AccountRequestAttachments. ex: ['employees', 'monthly']
-     * @param ids [list of strings, default null]: list of AccountRequestAttachment ids to filter retrieved objects. ex: ['5656565656565656', '4545454545454545']
+     * @param tags [list of strings, default null]: list of strings for reference when searching for IndividualAccountAttachments. ex: ['employees', 'monthly']
+     * @param ids [list of strings, default null]: list of IndividualAccountAttachment ids to filter retrieved objects. ex: ['5656565656565656', '4545454545454545']
      * @param user [Organization/Project object, default null]: Project object. Not necessary if starkinfra.user was set before function call
      *
      * Return:
-     * @returns list of AccountRequestAttachment objects with updated attributes
+     * @returns list of IndividualAccountAttachment objects with updated attributes
      *
      */
     let query = {
@@ -127,9 +127,9 @@ exports.query = async function ({ limit, after, before, status, tags, ids, user 
 exports.page = async function ({ cursor, limit, after, before, status, tags, ids, user } = {}) {
     /**
      *
-     * Retrieve paged AccountRequestAttachments
+     * Retrieve paged IndividualAccountAttachments
      *
-     * @description Receive a list of AccountRequestAttachment objects previously created in the Stark Infra API and the cursor to the next page.
+     * @description Receive a list of IndividualAccountAttachment objects previously created in the Stark Infra API and the cursor to the next page.
      *
      * Parameters (optional):
      * @param cursor [string, default null]: cursor returned on the previous page function call
@@ -137,12 +137,12 @@ exports.page = async function ({ cursor, limit, after, before, status, tags, ids
      * @param after [string, default null]: date filter for objects created after this date. ex: '2020-03-10'
      * @param before [string, default null]: date filter for objects created before this date. ex: '2020-03-10'
      * @param status [string, default null]: filter for status of the retrieved objects. ex: 'created', 'canceled', 'processing', 'failed', 'success'
-     * @param tags [list of strings, default null]: list of strings for reference when searching for AccountRequestAttachments. ex: ['employees', 'monthly']
-     * @param ids [list of strings, default null]: list of AccountRequestAttachment ids to filter retrieved objects. ex: ['5656565656565656', '4545454545454545']
+     * @param tags [list of strings, default null]: list of strings for reference when searching for IndividualAccountAttachments. ex: ['employees', 'monthly']
+     * @param ids [list of strings, default null]: list of IndividualAccountAttachment ids to filter retrieved objects. ex: ['5656565656565656', '4545454545454545']
      * @param user [Organization/Project object, default null]: Project object. Not necessary if starkinfra.user was set before function call
      *
      * Return:
-     * @returns list of AccountRequestAttachment objects with updated attributes and a cursor to the next page
+     * @returns list of IndividualAccountAttachment objects with updated attributes and a cursor to the next page
      *
      */
     let query = {
@@ -155,4 +155,24 @@ exports.page = async function ({ cursor, limit, after, before, status, tags, ids
         ids: ids
     };
     return rest.getPage(resource, query, user);
+};
+
+exports.delete = async function (id, { user } = {}) {
+    /**
+     *
+     * Delete an IndividualAccountAttachment entity
+     *
+     * @description Delete an IndividualAccountAttachment entity previously created in the Stark Infra API
+     *
+     * Parameters (required):
+     * @param id [string]: IndividualAccountAttachment unique id. ex: '5656565656565656'
+     *
+     * Parameters (optional):
+     * @param user [Organization/Project object]: Organization or Project object. Not necessary if starkinfra.user was set before function call
+     *
+     * Return:
+     * @returns deleted IndividualAccountAttachment object
+     *
+     */
+    return rest.deleteId(resource, id, user);
 };